*Reduced memory usage on recursive copy/move/delete/md5check. The memory is now freed after a process directly and not at the end of all the processes. This leaves more memory for further operations and so for further sub-folders. This should fix the crash on copy/move/delete/md5check of large directories and/or directories with many sub-folders. *Language files updated

+++ source/TextOperations/Text.cpp
@@ -125,7 +125,7 @@ void Text::SetMaxWidth(int w)
 
 void Text::SetTextLine(int line)
 {
-    curLineStart = cut_bounds(line, 0, TextLines.size()-1);
+    curLineStart = cut_bounds(line, 0, (int) TextLines.size()-1);
 
     FillRows();
